Scatterplot for several surfaces
It is now possible to insert a scatterplot for every level of surface (just for csv imported as it is not implemented for the img one yet)
for now you have to import csv as scatterplot. Now some surfaces have to be created. At the moment the level number where the points get insertet, hardcoded.